# src/main/resources/openapi/card-service.yaml
openapi: 3.0.3
info:
  title: Card Service API
  version: 1.0.0
  description: API for managing debit & credit cards
servers:
  - url: /api/v1
paths:
  /credit-cards:
    get:
      summary: Get all credit cards
      operationId: getAllCreditCards
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: string
          description: Filter credit cards by customer ID
          example: "68a0cdde7d68c910da0bc239"
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
          description: Filter credit cards by active status (default = true)
          example: true
      responses:
        '200':
          description: List of credit cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditCardResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new credit card
      operationId: createCreditCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardCreateRequest'
      responses:
        '201':
          description: Credit card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '400':
          description: Bad request - Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                personalCreditCardExists:
                  summary: Personal customer already has a credit card
                  value:
                    code: "PERSON_ALREADY_HAS_CREDIT_CARD"
                    message: "Personal customers can only have one active credit card"
                    timestamp: "2025-08-24T10:15:30Z"
                customerNotFound:
                  summary: Customer not found
                  value:
                    code: "CUSTOMER_NOT_FOUND"
                    message: "Customer with provided ID does not exist"
                    timestamp: "2025-08-24T10:15:30Z"
  /credit-cards/{id}:
    get:
      summary: Get a credit card by ID
      operationId: getCreditCardById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit card found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Credit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a credit card by ID
      operationId: updateCreditCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardUpdateRequest'
      responses:
        '200':
          description: Credit card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Credit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request - Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a credit card by ID
      operationId: deleteCreditCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Credit card deleted
        '404':
          description: Credit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit-cards/{id}/deactivate:
    patch:
      summary: Soft delete (deactivate) a credit card by ID
      description: Marks the credit card as inactive without physically removing it
      operationId: deactivateCreditCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit card deactivated (isActive set to false)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Credit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit-cards/{id}/activate:
    patch:
      summary: Reactivate a credit card by ID
      description: Marks the credit card as active again
      operationId: activateCreditCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit card activated (isActive set to true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Credit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit-cards/customers/{customerId}/has-active-card:
    get:
      summary: Check if customer has an active credit card
      description: Validates if a customer has at least one active credit card for account eligibility requirements
      operationId: checkCustomerHasActiveCard
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
          description: Customer ID to check for active credit cards
          example: "68a0cdde7d68c910da0bc239"
      responses:
        '200':
          description: Customer credit card validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCardValidationResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit-cards/{cardNumber}/authorize-charge:
    patch:
      summary: Authorize credit card charge
      description: Decreases the availableCredit and increases the currentBalance if enough credit is available.
      operationId: authorizeCharge
      parameters:
        - name: cardNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargeAuthorizationRequest'
      responses:
        '200':
          description: Credit reserved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargeAuthorizationResponse'
        '400':
          description: Insufficient credit or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit-cards/{cardNumber}/process-payment:
    post:
      summary: Process payment validation and update balances
      description: Validates payment and updates card balances if successful
      operationId: processCardPayment
      parameters:
        - name: cardNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentProcessRequest'
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProcessResponse'
        '400':
          description: Payment validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit-cards/{cardNumber}/balance:
    get:
      summary: Get credit card balance information
      operationId: getCardBalance
      parameters:
        - name: cardNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Card balance information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardBalanceResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /credit-cards/customers/{customerId}/daily-averages:
    get:
      summary: Get customer's credit cards daily average balances
      operationId: getCustomerDailyAverages
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
        - name: year
          in: query
          required: true
          schema:
            type: integer
          example: 2025
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
          example: 8
      responses:
        '200':
          description: Daily balance average
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDailyAverageResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /credit-cards/customers/{customerId}/product-eligibility:
    get:
      summary: Check customer eligibility for new products
      description: Validates if customer can acquire new products (no overdue credit debts)
      operationId: checkCustomerProductEligibility
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Eligibility status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductEligibilityResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CreditCardCreateRequest:
      type: object
      properties:
        customerId:
          type: string
          example: "68a0cdde7d68c910da0bc239"
        creditLimit:
          type: number
          format: double
          example: 10000.00
        availableCredit:
          type: number
          format: double
          example: 8000
        currentBalance:
          type: number
          format: double
          example: 2000
      required:
        - creditLimit
    CreditCardUpdateRequest:
      type: object
      properties:
        creditLimit:
          type: number
          format: double
          example: 12000.00
        availableCredit:
          type: number
          format: double
          example: 8000
        currentBalance:
          type: number
          format: double
          example: 4000
        isActive:
          type: boolean
          example: true
      required:
        - creditLimit
        - availableCredit
        - currentBalance
        - isActive
    ChargeAuthorizationRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Amount to authorize
          example: 150.00
...
    PaymentProcessRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Payment amount
          example: 500.00
    PaymentProcessResponse:
      type: object
      required:
        - success
        - cardId
        - requestedAmount
        - processedAt
      properties:
        success:
          type: boolean
          description: Whether the payment was successful
          example: true
        cardId:
          type: string
          description: Card ID
          example: "64f3cbb5f8a1e53d2f7c34e1"
        requestedAmount:
          type: number
          format: double
          description: Requested payment amount
          example: 500.00
        actualPaymentAmount:
          type: number
          format: double
          description: Actual amount paid (only if success = true)
          example: 500.00
        availableCreditAfter:
          type: number
          format: double
          description: Available credit after payment (only if success = true)
          example: 9500.00
        currentBalanceAfter:
          type: number
          format: double
          description: Current balance after payment (only if success = true)
          example: 500.00
        errorCode:
          type: string
          enum: [ CARD_INACTIVE, INVALID_AMOUNT, OVERPAYMENT_NOT_ALLOWED, ZERO_CURRENT_BALANCE ]
          description: Error code (only if success = false)
          example: "OVERPAYMENT_NOT_ALLOWED"
        errorMessage:
          type: string
          description: Error message (only if success = false)
          example: "Payment amount exceeds current balance"
        processedAt:
          type: string
          format: date-time
          description: Processing timestamp
          example: "2025-08-30T15:30:00Z"
    ChargeAuthorizationResponse:
      type: object
      required:
        - status
        - authorizedAmount
        - availableCreditAfter
        - processedAt
      properties:
        authorizationCode:
          type: string
          description: Authorization code (null if it was rejected)
          example: "AUTH-456789"
        status:
          type: string
          enum: [ APPROVED, DECLINED ]
          description: Authorization status
          example: "APPROVED"
        authorizedAmount:
          type: number
          format: double
          description: Authorized amount (0 if it was rejected)
          example: 150.00
        availableCreditAfter:
          type: number
          format: double
          description: Available credit after the transaction
          example: 850.00
        processedAt:
          type: string
          format: date-time
          description: Processing Timestamp
          example: "2025-08-28T10:30:05Z"
        declineReason:
          type: string
          enum: [ INSUFFICIENT_CREDIT, CARD_INACTIVE ]
          description: Reject reason (only if status = DECLINED)
          example: "INSUFFICIENT_CREDIT"
    CreditCardResponse:
      type: object
      properties:
        id:
          type: string
          example: "64f3cbb5f8a1e53d2f7c34e1"
        cardNumber:
          type: string
          example: "1234-5678-9123-4567"
        customerId:
          type: string
          example: "68a0cdde7d68c910da0bc239"
        creditCardType:
          type: string
          enum: [ "PERSONAL", "ENTERPRISE" ]
          example: "PERSONAL"
        creditLimit:
          type: number
          format: double
          example: 10000.00
        availableCredit:
          type: number
          format: double
          example: 6000
        currentBalance:
          type: number
          format: double
          example: 4000
        isActive:
          type: boolean
          example: true
        paymentDueDate:
          type: string
          format: date
          example: "2025-09-15"
          description: "Payment due date"
        minimumPayment:
          type: number
          format: double
          example: 200.00
          description: "Minimum required payment"
        isOverdue:
          type: boolean
          example: false
          description: "Indicates whether the card has overdue payments"
        overdueDays:
          type: integer
          example: 0
          description: "Number of overdue days"
        createdAt:
          type: string
          format: date-time
          example: 2025-08-18T10:15:30Z
        updatedAt:
          type: string
          format: date-time
          example: 2025-08-18T10:15:30Z
    CreditCardBalanceResponse:
      type: object
      required:
        - cardId
        - cardNumber
        - creditLimit
        - availableCredit
        - currentBalance
        - utilizationPercentage
      properties:
        cardId:
          type: string
          description: Card ID
          example: "64f3cbb5f8a1e53d2f7c34e1"
        cardNumber:
          type: string
          description: Card number
          example: "1234567891234567"
        creditLimit:
          type: number
          format: double
          description: Total credit limit
          example: 10000.00
        availableCredit:
          type: number
          format: double
          description: Available credit
          example: 6000.00
        currentBalance:
          type: number
          format: double
          description: Current balance (debt)
          example: 4000.00
        utilizationPercentage:
          type: number
          format: double
          description: Credit utilization percentage
          example: 40.00
        isActive:
          type: boolean
          description: Card active status
          example: true
    CustomerCardValidationResponse:
      type: object
      required:
        - customerId
        - hasActiveCard
        - activeCardCount
        - validatedAt
      properties:
        customerId:
          type: string
          description: Customer ID that was validated
          example: "68a0cdde7d68c910da0bc239"
        hasActiveCard:
          type: boolean
          description: Whether the customer has at least one active credit card
          example: true
        activeCardCount:
          type: integer
          description: Total number of active credit cards for this customer
          example: 1
        cardSummary:
          type: array
          description: Summary of active cards (optional, only if hasActiveCard = true)
          items:
            type: object
            properties:
              cardId:
                type: string
                example: "64f3cbb5f8a1e53d2f7c34e1"
              cardNumber:
                type: string
                example: "1234-5678-9123-4567"
              type:
                type: string
                enum: [ "CREDIT", "DEBIT" ]
                example: "CREDIT"
              creditLimit:
                type: number
                format: double
                example: 10000.00
              availableCredit:
                type: number
                format: double
                example: 6000.00
        validatedAt:
          type: string
          format: date-time
          description: Timestamp when the validation was performed
          example: "2025-08-31T15:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "PERSON_ALREADY_HAS_CREDIT_CARD"
        message:
          type: string
          example: "Personal customers can only have one active credit card"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-24T10:15:30Z"
      required:
        - code
        - message
        - timestamp